fix build warning in valijson
Fix build time warning when building valijson for 32-bit. This has
already been fixed upstream, but avoid reimporting the new headers
as valijson seem to work fine otherwise.

  valijson/constraints/concrete_constraints.hpp:481: warning: conversion
  from ‘long long unsigned int’ to ‘size_t’ {aka ‘unsigned int’} changes
  value from ‘18446744073709551615’ to ‘4294967295’ [-Woverflow]
           maxProperties(std::numeric_limits<uint64_t>::max()) { }
                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~

+++ valijson/constraints/concrete_constraints.hpp
@@ -437,7 +437,7 @@ public:
     }
 
 private:
-    size_t maxItems;
+    uint64_t maxItems;
 };
 
 /**
@@ -491,7 +491,7 @@ public:
     }
 
 private:
-    size_t maxProperties;
+    uint64_t maxProperties;
 };
 
 /**
@@ -558,7 +558,7 @@ public:
     }
 
 private:
-    size_t minItems;
+    uint64_t minItems;
 };
 
 /**
